requirements.txt
streamlit
torch
torchvision
timm
gdown
Pillow